pyproject.toml
[tool.poetry]
name = "sophosfirewall-ansible"
version = "1.2.1"
description = "Ansible modules for Sophos Firewall"
authors = ["Matt Mullen <matt.mullen@sophos.com>"]
license = "GNU GENERAL PUBLIC LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sophosfirewall-python = "^0.1.56"
ansible-base = "^2.10.17"

[tool.poetry.group.dev.dependencies]
antsibull = "^0.67.0"
antsibull-docs = "^2.15.0"
black = "24.10.0"
pylint = "^3.3.1"
sphinx-rtd-theme = "^3.0.1"
sphinx-ansible-theme = "^0.10.3"
Sphinx = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
